+++ docs/src/lib/registry/blocks/sidebar-16/components/site-header.svelte
@@ -18,7 +18,7 @@
 		<Breadcrumb.Root class="hidden sm:block">
 			<Breadcrumb.List>
 				<Breadcrumb.Item>
-					<Breadcrumb.Link href="#">Building Your Application</Breadcrumb.Link>
+					<Breadcrumb.Link href="##">Building Your Application</Breadcrumb.Link>
 				</Breadcrumb.Item>
 				<Breadcrumb.Separator />
 				<Breadcrumb.Item>